requirements.txt
langchain
langchain-groq
langchain-huggingface
langchain-text-splitters
langchain-community
chromadb
torch
sentence-transformers
python-dotenv
